refactor(gateway): extract startup bootstrap module
Move pre-loop startup responsibilities into bootstrapGateway so index focuses on launching runtime loops and process fatal handling.
